mysql-versionup-test/mysql-compare/python/requirements.txt
pytest
mysql-connector-python
PyMySQL
pytest-json-report
